test SDK 1112

+++ src/main/java/com/cobo/waas2/model/SwapActivity.java
@@ -116,11 +116,11 @@ public class SwapActivity {
 
   public static final String SERIALIZED_NAME_CREATED_TIMESTAMP = "created_timestamp";
   @SerializedName(SERIALIZED_NAME_CREATED_TIMESTAMP)
-  private Integer createdTimestamp;
+  private Long createdTimestamp;
 
   public static final String SERIALIZED_NAME_UPDATED_TIMESTAMP = "updated_timestamp";
   @SerializedName(SERIALIZED_NAME_UPDATED_TIMESTAMP)
-  private Integer updatedTimestamp;
+  private Long updatedTimestamp;
 
   public static final String SERIALIZED_NAME_NETWORK_FEE = "network_fee";
   @SerializedName(SERIALIZED_NAME_NETWORK_FEE)
@@ -196,7 +196,7 @@ public class SwapActivity {
   }
 
    /**
-   * The request ID of the swap transaction.
+   * The request id of the swap activity.
    * @return requestId
   **/
   @javax.annotation.Nullable
@@ -215,7 +215,7 @@ public class SwapActivity {
   }
 
    /**
-   * The ID of the wallet used to pay.
+   * The unique identifier of the wallet.
    * @return walletId
   **/
   @javax.annotation.Nullable
@@ -234,7 +234,7 @@ public class SwapActivity {
   }
 
    /**
-   * The ID of the token to pay.
+   * The token ID to pay.
    * @return payTokenId
   **/
   @javax.annotation.Nullable
@@ -253,7 +253,7 @@ public class SwapActivity {
   }
 
    /**
-   * The ID of the token to receive.
+   * The token ID to receive.
    * @return receiveTokenId
   **/
   @javax.annotation.Nullable
@@ -272,7 +272,7 @@ public class SwapActivity {
   }
 
    /**
-   * The amount of the token to pay.
+   * The amount of tokens to bridge.
    * @return payAmount
   **/
   @javax.annotation.Nullable
@@ -291,7 +291,7 @@ public class SwapActivity {
   }
 
    /**
-   * The amount of the token to receive.
+   * The amount of tokens to receive.
    * @return receiveAmount
   **/
   @javax.annotation.Nullable
@@ -310,7 +310,7 @@ public class SwapActivity {
   }
 
    /**
-   * The ID of the token used for paying the service fee.
+   * The fee token ID.
    * @return feeTokenId
   **/
   @javax.annotation.Nullable
@@ -329,7 +329,7 @@ public class SwapActivity {
   }
 
    /**
-   * The amount of the service fee.
+   * The amount of fee.
    * @return feeAmount
   **/
   @javax.annotation.Nullable
@@ -399,7 +399,7 @@ public class SwapActivity {
   }
 
 
-  public SwapActivity createdTimestamp(Integer createdTimestamp) {
+  public SwapActivity createdTimestamp(Long createdTimestamp) {
     this.createdTimestamp = createdTimestamp;
     return this;
   }
@@ -409,16 +409,16 @@ public class SwapActivity {
    * @return createdTimestamp
   **/
   @javax.annotation.Nullable
-  public Integer getCreatedTimestamp() {
+  public Long getCreatedTimestamp() {
     return createdTimestamp;
   }
 
-  public void setCreatedTimestamp(Integer createdTimestamp) {
+  public void setCreatedTimestamp(Long createdTimestamp) {
     this.createdTimestamp = createdTimestamp;
   }
 
 
-  public SwapActivity updatedTimestamp(Integer updatedTimestamp) {
+  public SwapActivity updatedTimestamp(Long updatedTimestamp) {
     this.updatedTimestamp = updatedTimestamp;
     return this;
   }
@@ -428,11 +428,11 @@ public class SwapActivity {
    * @return updatedTimestamp
   **/
   @javax.annotation.Nullable
-  public Integer getUpdatedTimestamp() {
+  public Long getUpdatedTimestamp() {
     return updatedTimestamp;
   }
 
-  public void setUpdatedTimestamp(Integer updatedTimestamp) {
+  public void setUpdatedTimestamp(Long updatedTimestamp) {
     this.updatedTimestamp = updatedTimestamp;
   }
 
@@ -462,7 +462,7 @@ public class SwapActivity {
   }
 
    /**
-   * The address of an MPC Wallet or Web3 Wallet that receives the swapped or bridged assets.
+   * the destination address of web3/mpc wallets.
    * @return destinationAddress
   **/
   @javax.annotation.Nullable
