+++ src/main/java/com/cobo/waas2/model/SwapActivityTimeline.java
@@ -62,7 +62,7 @@ public class SwapActivityTimeline {
 
   public static final String SERIALIZED_NAME_TIMESTAMP = "timestamp";
   @SerializedName(SERIALIZED_NAME_TIMESTAMP)
-  private Integer timestamp;
+  private Long timestamp;
 
   public SwapActivityTimeline() {
   }
@@ -73,7 +73,7 @@ public class SwapActivityTimeline {
   }
 
    /**
-   * The action in the swap activity. Possible values include:   - &#x60;Submitted&#x60;: The swap request has been submitted.   - &#x60;Pending Authorization&#x60;: The swap request is pending authorization.   - &#x60;Bridge {Token}&#x60;: The token is being bridged to the destination chain.   - &#x60;Swap {Token}&#x60;: The token is being exchanged to the destination token.   - &#x60;Cobo Confirmation&#x60;: The swap result is pending Cobo&#39;s final confirmation. 
+   * The action of the swap activity. Possible values include:   - &#x60;Submitted&#x60;: The swap request is submitted.   - &#x60;Pending Authorization&#x60;: The swap request is pending authorization.   - &#x60;Bridge {Token}&#x60;: The token is being bridged to the target chain.   - &#x60;Swap {Token}&#x60;: The token is being swapped on the target chain.   - &#x60;Cobo Confirmation&#x60;: The swap result is waiting for Cobo confirmation. 
    * @return action
   **/
   @javax.annotation.Nonnull
@@ -92,7 +92,7 @@ public class SwapActivityTimeline {
   }
 
    /**
-   * The status of the action. Possible values include:   - &#x60;Success&#x60;: The action completed successfully.   - &#x60;Processing&#x60;: The action is being processed.   - &#x60;Failed&#x60;: The action failed. 
+   * The status of the action. Possible values include:   - &#x60;Success&#x60;: The action is successfully completed.   - &#x60;Processing&#x60;: The action is being processed.   - &#x60;Failed&#x60;: The action has failed. 
    * @return status
   **/
   @javax.annotation.Nonnull
@@ -105,21 +105,21 @@ public class SwapActivityTimeline {
   }
 
 
-  public SwapActivityTimeline timestamp(Integer timestamp) {
+  public SwapActivityTimeline timestamp(Long timestamp) {
     this.timestamp = timestamp;
     return this;
   }
 
    /**
-   * The time when the action occurred, in Unix timestamp format, measured in milliseconds.  
+   * The time when the action took place, in Unix timestamp format, measured in milliseconds.  
    * @return timestamp
   **/
   @javax.annotation.Nullable
-  public Integer getTimestamp() {
+  public Long getTimestamp() {
     return timestamp;
   }
 
-  public void setTimestamp(Integer timestamp) {
+  public void setTimestamp(Long timestamp) {
     this.timestamp = timestamp;
   }
